Increase maximum phrase length to 64 quarter notes.
The documentation I read stated there was a 16 quarter note limit to the
phrase length, but that is not true.  The vanilla battle theme phrases
three and four are longer than four bars.  Presumably there is actually
no limit to the length of a phrase other than the number of notes you
can pack into 256 bytes because of the indexing.