Fix how editors use commits.
Previously, editor widgets would maintain their own nottion of the
commit index of an edit and use it when updating an edit.  However,
since edits can be reordered, the commit index is not a stable
identifier.

Instead, editors simply keep a copy of the `Edit` struct and commit
that to the project.  The project will figure out the relevant indicies
and replay the necessary commits.