Write the drop table correctly.
Fixes a bug found by OkImpala: the drop table was not being written with
the MSB set, therefore causing drops to affect the players item
availability bitmap.  The bug would manifest as such that  taking a drop from
an enemy could cause the game to mark a fixed item as taken.